Move addresses to own package
...to avoid circular dependencies.
