+++ core/vm/celo_contracts.go
@@ -6,11 +6,10 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/common/math"
+	"github.com/ethereum/go-ethereum/contracts/addresses"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/holiman/uint256"
 )
-
-var GoldTokenAddress = common.HexToAddress("0x471ece3750da237f93b8e339c536989b8978a438")
 
 type CeloPrecompiledContract interface {
 	RequiredGas(input []byte) uint64                              // RequiredGas calculates the contract gas use
@@ -48,7 +47,7 @@ func celoPrecompileAddress(index byte) common.Address {
 }
 
 func (ctx *celoPrecompileContext) IsCallerGoldToken() (bool, error) {
-	return GoldTokenAddress == ctx.caller, nil
+	return addresses.GoldTokenAddress == ctx.caller, nil
 }
 
 // Native transfer contract to make Celo Gold ERC20 compatible.
